pyproject.toml
[project]
name = "kokkai-scorer"
version = "0.1.0"
description = "国会審議品質スコアボード — AI評価による国会質疑の可視化"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.31",
    "anthropic>=0.40",
    "tweepy>=4.14",
    "playwright>=1.40",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[project.scripts]
kokkai-scorer = "daily_pipeline:main"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
